% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_crs}
\alias{check_crs}
\title{Check CRS of an sf Object}
\usage{
check_crs(sf_object)
}
\arguments{
\item{sf_object}{The 'sf' object to check, representing geographical features
and their attributes.}
}
\value{
Does not return a value but halts execution with an error if the CRS
of 'sf_object' is not EPSG:4326.
}
\description{
Verifies that the coordinate reference system (CRS) of an 'sf' object is set
to EPSG:4326, which is required for specific geographical applications. This
CRS corresponds to WGS 84, a global geographic coordinate system. The
function halts execution with an error if the CRS is not EPSG:4326.
}
\keyword{internal}
