% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ecs_input_measurements.R
\name{create_ecs_input_measurements}
\alias{create_ecs_input_measurements}
\title{Create ECS Fault Displacement Measurements Input File from Esri Shapefile}
\usage{
create_ecs_input_measurements(in_filepath, output_folder = NULL)
}
\arguments{
\item{in_filepath}{The path to the input Esri shapefile, recommended to use
'file.path()' for construction.}

\item{output_folder}{(Optional) Directory for saving the output CSV file;
defaults to 'ECSprep-outputs' in the input file directory.}
}
\value{
Outputs a CSV file to the specified directory and prints the file
location. Does not return any value from the function itself.
}
\description{
Processes an Esri shapefile with point geometry for input to the Lavrentiadis
& Abrahamson ECS tool. The function transforms the shapefile's CRS to
EPSG:4326, adds a default 'Rank' field if missing, and formats the data for
ECS tool input. The output is saved as a CSV file in the specified or default
directory. Note that a column called 'displ' that contains the fault
displacement measurements is required and execution is halted with an error
message if it is not present.
}
