% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_hazsite}
\alias{process_hazsite}
\title{Convert hazard analysis sites to a data frame with necessary attributes for
the ECS Tool}
\usage{
process_hazsite(sf_object)
}
\arguments{
\item{sf_object}{An 'sf' object, expected to contain POINT geometries
representing hazard evaluation sites.}
}
\value{
Returns a data frame derived from the input 'sf' object. The data
frame includes the original data and coordinates ('Latitude', 'Longitude'
in EPSG:4326).
}
\description{
Transforms an 'sf' object containing POINT geometries into a data frame. This
function is tailored for preparing input data for the Lavrentiadis &
Abrahamson ECS tool .
}
\keyword{internal}
