% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_points}
\alias{process_points}
\title{Convert measurement sites to format for the Lavrentiadis & Abrahamson ECS tool.}
\usage{
process_points(sf_object)
}
\arguments{
\item{sf_object}{The sf object with POINT geometry.}
}
\value{
A data frame with the points .
}
\description{
Convert measurement sites to format for the Lavrentiadis & Abrahamson ECS tool.
}
\keyword{internal}
