% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rups2verts.R
\name{rups2verts}
\alias{rups2verts}
\title{Convert rupture line work to vertex format for the Lavrentiadis & Abrahamson ECS tool.}
\usage{
rups2verts(sf_object)
}
\arguments{
\item{sf_object}{The sf object with LINESTRING geometry.}
}
\value{
A data frame with the rupture lines converted to vertices.
}
\description{
Convert rupture line work to vertex format for the Lavrentiadis & Abrahamson ECS tool.
}
