% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_crs}
\alias{check_crs}
\title{Check CRS of an sf Object}
\usage{
check_crs(sf_object)
}
\arguments{
\item{sf_object}{The 'sf' object to check, representing geographical features
and their attributes.}
}
\value{
Does not return a value but halts execution with an error if there is
no CRS for the 'sf_object' .
}
\description{
Verifies that a coordinate reference system (CRS) is provided for the 'sf'
object, which is required for specific geographical applications. This The
function halts execution with an error if the CRS is not set.
}
\keyword{internal}
