% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_crs}
\alias{check_crs}
\title{Check the coordinate reference system of an sf object. It should be 4326 for
our application.}
\usage{
check_crs(sf_object)
}
\arguments{
\item{sf_object}{The sf object.}
}
\value{
None; the function will stop execution and throw an error if the CRS
of sf_object is not 4326.
}
\description{
Check the coordinate reference system of an sf object. It should be 4326 for
our application.
}
