% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_geometry_type}
\alias{check_geometry_type}
\title{Check the geometry of an sf object.}
\usage{
check_geometry_type(sf_object, geom_category)
}
\arguments{
\item{sf_object}{The sf object.}

\item{geom_category}{A string describing the expected geometry type; only
'line' or 'point' are allowed.}
}
\value{
None; the function will stop execution and throw an error if the
geometry type of sf_object does not match the expected type.
}
\description{
Check the geometry of an sf object.
}
