% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ranking.R
\name{check_ranking}
\alias{check_ranking}
\title{Check that the values in the Rank attribute are valid; allowable values are
None, Primary, Principal, Secondary, Distributed. Note that Principal and
Distributed will be replaced with Primary and Secondary, respectively, for
standardization purposes.}
\usage{
check_ranking(sf_object)
}
\arguments{
\item{sf_object}{The sf object.}
}
\value{
The sf object with rank cleaned-up as needed. The function will stop
execution and throw an error if an invalid Rank is found.
}
\description{
Check that the values in the Rank attribute are valid; allowable values are
None, Primary, Principal, Secondary, Distributed. Note that Principal and
Distributed will be replaced with Primary and Secondary, respectively, for
standardization purposes.
}
