% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{project_to_4326}
\alias{project_to_4326}
\title{Check the coordinate reference system of an sf object; it should be
EPSG:4326; covert if needed.}
\usage{
project_to_4326(sf_object)
}
\arguments{
\item{sf_object}{The sf object.}
}
\value{
An sf object in EPSG:4326.
}
\description{
Check the coordinate reference system of an sf object; it should be
EPSG:4326; covert if needed.
}
