% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rups2verts}
\alias{rups2verts}
\title{Convert rupture line work to a data frame of vertices with necessary
attributes for the ECS Tool}
\usage{
rups2verts(sf_object)
}
\arguments{
\item{sf_object}{An 'sf' object, expected to contain LINESTRING geometries.}
}
\value{
Returns a data frame where each row represents a vertex from the
original LINESTRING geometries. The data frame includes columns for RUP_ID
(rupture line ID), NODE_ID (node order within each RUP_ID), and
Latitude/Longitude coordinates.
}
\description{
Transforms an 'sf' object containing LINESTRING geometries into a data frame
of vertices. Each vertex is assigned a unique node ID within its rupture line
(RUP_ID) and includes latitude and longitude coordinates. This function is
tailored for preparing input data for the Lavrentiadis & Abrahamson ECS tool
by structuring the 'sf' object's line work into a vertex format in a data
frame.
}
\keyword{internal}
